✨Feat: 컴포넌트 분리 및 데이터 연동
