+++ src/app/(main)/main/page.tsx
@@ -10,35 +10,62 @@ import { getUserProfile } from "@/lib/api/user";
 export const dynamic = "force-dynamic";
 export const revalidate = 0;
 
+// Default screen time data structure
+const defaultScreenTimeData = {
+  success: true,
+  message: "Using default screen time data",
+  data: {
+    screenTimes: [{
+      date: new Date().toISOString().split("T")[0],
+      totalMinutes: 0,
+      appTimes: {
+        instagram: 0,
+        youtube: 0,
+        kakaotalk: 0,
+        chrome: 0
+      }
+    }]
+  }
+};
+
 export default async function HomePage() {
   let userProfile = null;
-  let screenTimeData = null;
+  let screenTimeData = defaultScreenTimeData;
 
-  // 1) 사용자 프로필 조회와 스크린타임 생성/갱신을 병렬 수행
-  const [profileSettled, generateSettled] = await Promise.allSettled([
-    getUserProfile(),
-    createOrUpdateScreenTime(),
-  ]);
-
-  if (profileSettled.status === "fulfilled") {
-    userProfile = profileSettled.value;
-    console.log("🔍 Main 페이지: 사용자 프로필 조회 결과(서버):", userProfile);
-  } else {
-    console.error("메인: 사용자 프로필 조회 실패:", profileSettled.reason);
-  }
-
-  if (generateSettled.status === "fulfilled") {
-    console.log("🆗 스크린타임 생성/갱신 결과:", generateSettled.value);
-  } else {
-    console.error("메인: 스크린타임 생성/갱신 실패:", generateSettled.reason);
-  }
-
-  // 2) 생성/갱신 이후, 일간 스크린타임 조회
   try {
-    screenTimeData = await getScreenTimeDay();
-    console.log("🔍 Main 페이지: 스크린타임 조회 결과(서버):", screenTimeData);
-  } catch (e) {
-    console.error("메인: 스크린타임 조회 실패:", e);
+    // 1) 사용자 프로필 조회와 스크린타임 생성/갱신을 병렬 수행
+    const [profileResponse, screenTimeResponse] = await Promise.allSettled([
+      getUserProfile(),
+      createOrUpdateScreenTime()
+    ]);
+
+    // Handle user profile response
+    if (profileResponse.status === "fulfilled") {
+      userProfile = profileResponse.value;
+      console.log("🔍 Main 페이지: 사용자 프로필 조회 결과(서버):", userProfile);
+    } else {
+      console.error("메인: 사용자 프로필 조회 실패:", profileResponse.reason);
+    }
+
+    // Handle screen time response
+    if (screenTimeResponse.status === "fulfilled") {
+      console.log("🆗 스크린타임 생성/갱신 결과:", screenTimeResponse.value);
+      
+      // 2) 생성/갱신 이후, 일간 스크린타임 조회
+      try {
+        const dayData = await getScreenTimeDay();
+        if (dayData.success && dayData.data?.screenTimes?.[0]) {
+          screenTimeData = dayData;
+        }
+        console.log("🔍 Main 페이지: 스크린타임 조회 결과(서버):", screenTimeData);
+      } catch (e) {
+        console.error("메인: 스크린타임 조회 실패:", e);
+      }
+    } else {
+      console.error("메인: 스크린타임 생성/갱신 실패:", screenTimeResponse.reason);
+    }
+  } catch (error) {
+    console.error("메인 페이지 초기화 중 오류 발생:", error);
   }
 
   return (
