+++ src/lib/api/screentime.ts
@@ -23,10 +23,52 @@ export const getScreenTimeDay = async (
   date?: string
 ): Promise<ScreenTimeResponse> => {
   const today = date || new Date().toISOString().split("T")[0];
-  const { data } = await privateApi.get<ScreenTimeResponse>(
-    `/api/screentime?period=day&date=${today}`
-  );
-  return data;
+  try {
+    const { data } = await privateApi.get<ScreenTimeResponse>(
+      `/api/screentime?period=day&date=${today}`
+    );
+    
+    // If screenTimes is empty or null, provide default values
+    if (!data.data?.screenTimes?.[0]) {
+      return {
+        success: true,
+        message: "No screen time data available",
+        data: {
+          screenTimes: [{
+            date: today,
+            totalMinutes: 0,
+            appTimes: {
+              instagram: 0,
+              youtube: 0,
+              kakaotalk: 0,
+              chrome: 0
+            }
+          }]
+        }
+      };
+    }
+    
+    return data;
+  } catch (error) {
+    console.error("Error fetching screen time:", error);
+    // Return default data structure on error
+    return {
+      success: false,
+      message: "Failed to fetch screen time data",
+      data: {
+        screenTimes: [{
+          date: today,
+          totalMinutes: 0,
+          appTimes: {
+            instagram: 0,
+            youtube: 0,
+            kakaotalk: 0,
+            chrome: 0
+          }
+        }]
+      }
+    };
+  }
 };
 
 // 주간 스크린타임 조회
